AspDotNetCoreFromScratch - Angular ClientApp - connected checkout to server-side Orders.PostNewOrder.
Note: one of the changes I had to do to avoid DB exception when adding the new order was not setting the itemId in addItemToOrder.
This broke the code recognizing the the item already exist in the order and update its quantity.
Since this course's code is a bit of a mess right now and I don't have time to go into it, I'm leaving this bug for now.